refactor(qvism): generate recipe-shared module from cli
* Ensure reproducible generated result
